refactoring and add comments
